refactor: extract logging + counters into processor/output module
Move 13 log_* formatting functions and RequestCounters/TotalCounters
from processor/executor.rs into a dedicated processor/output.rs module.

No behavioral changes — all 987 existing tests pass.
